requirements.txt
Flask
Flask-CORS
python-dotenv
requests
groq
elevenlabs
transformers
torch
torchvision
torchaudio
sentencepiece